requirements.txt
Pillow>=10.0.0
PyPDF2>=3.0.0
pypdf>=4.0.0
pillow-heif>=0.16.0